DOA COMPLETION

+++ 07_Data_Access_and_DAO/exercise/src/main/java/com/techelevator/projects/model/Department.java
@@ -2,12 +2,12 @@ package com.techelevator.projects.model;
 
 public class Department {
 
-	private Long id;
+	private int id;
 	private String name;
 
 	public Department() {}
 
-	public Department(Long id, String name) {
+	public Department(int id, String name) {
 		this.id = id;
 		this.name = name;
 	}
@@ -20,11 +20,11 @@ public class Department {
 		this.name = name;
 	}
 
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
 	
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
